Former-commit-id: b4bb9952616bcf5272dee690aa1cf0ee42e2d476

+++ libhpx/scheduler/netfuture.c
@@ -67,7 +67,7 @@ typedef struct {
 /// table at each locality.
 typedef struct {
   int table_index;
-  int offset; // from base
+  size_t offset; // from base
 } _fut_info_t;
 
 /// This struct manages the data the entire netfutures system at this
